Add a common ReactiveProducerConsumerStreamObserver abstract class
